fix some permission check
